--- a/arch/riscv/Kconfig
+++ b/arch/riscv/Kconfig
@@ -97,6 +97,7 @@ config RISCV
 	select SYSCTL_EXCEPTION_TRACE
 	select THREAD_INFO_IN_TASK
 	select UACCESS_MEMCPY if !MMU
+	select HANDLE_DOMAIN_IRQ # FIXME: should go in Kconfigs.soc !!!
 
 config ARCH_MMAP_RND_BITS_MIN
 	default 18 if 64BIT