--- a/arch/riscv/configs/litex_rocket_initramfs.config
+++ b/arch/riscv/configs/litex_rocket_initramfs.config
@@ -1,0 +1,1 @@
+CONFIG_INITRAMFS_SOURCE="initramfs.cpio"