--- a/drivers/gpu/drm/litevideo/Kconfig
+++ b/drivers/gpu/drm/litevideo/Kconfig
@@ -1,0 +1,18 @@
+# SPDX-License-Identifier: GPL-2.0
+
+# Copyright (C) 2020 Antmicro <www.antmicro.com>
+#
+# This program is free software; you can redistribute it and/or modify
+# it under the terms of the GNU General Public License version 2
+# as published by the Free Software Foundation.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+
+config DRM_LITEVIDEO
+	tristate "LiteVideo support"
+	depends on DRM && LITEX_SOC_CONTROLLER
+	help
+	  This enables DRM driver for LiteX LiteVideo.