--- a/drivers/gpu/drm/litevideo/Makefile
+++ b/drivers/gpu/drm/litevideo/Makefile
@@ -1,0 +1,3 @@
+# SPDX-License-Identifier: GPL-2.0
+
+obj-$(CONFIG_DRM_LITEVIDEO) += litevideo.o