--- a/drivers/hwmon/Kconfig
+++ b/drivers/hwmon/Kconfig
@@ -1180,6 +1180,12 @@ config SENSORS_ADCXX
 	  This driver can also be built as a module. If so, the module
 	  will be called adcxx.
 
+config SENSORS_LITEX_HWMON
+	tristate "LiteX XADC hardware monitor"
+	depends on OF && HAS_IOMEM && LITEX_SOC_CONTROLLER
+	help
+	  This enables hardware monitoring using LiteX XADC.
+
 config SENSORS_LM63
 	tristate "National Semiconductor LM63 and compatibles"
 	depends on I2C