… Documentation/devicetree/bindings/clock/litex,clock.yaml
… Documentation/devicetree/bindings/fpga/litex-fpga.txt
… Documentation/devicetree/bindings/gpio/litex,gpio.txt
… Documentation/devicetree/bindings/gpu/litex,litevideo.txt
… Documentation/devicetree/bindings/hwmon/litex-hwmon.txt
… Documentation/devicetree/bindings/i2c/i2c-litex.txt
… Documentation/devicetree/bindings/mtd/litex,spiflash.yaml
… Documentation/devicetree/bindings/pwm/litex,pwm.txt
… arch/riscv/configs/litex_rocket_defconfig
… drivers/gpu/drm/litevideo/litevideo.h
… drivers/hwmon/litex-hwmon.c
… drivers/i2c/busses/Kconfig
… drivers/i2c/busses/Makefile
… drivers/i2c/busses/i2c-litex.c
… drivers/irqchip/Kconfig
… drivers/irqchip/Makefile
… drivers/irqchip/irq-litex-vexriscv.c
… drivers/mmc/host/litex_mmc.c
… drivers/mmc/host/litex_mmc.h
… drivers/mtd/spi-nor/Kconfig
… drivers/mtd/spi-nor/Makefile
… drivers/mtd/spi-nor/litex-spiflash.c
… drivers/pwm/Kconfig
… drivers/pwm/Makefile
… drivers/pwm/pwm-litex.c