This adds support to the litex_mmc driver for using interrupts and for waiting while the card indicates it is busy.

There are a few other minor improvements; the MMC subsystem notices that the card has been removed much more quickly now, and the driver is less verbose in the amount it prints to the console.

Some changes to the device-tree node for the interface are required if you want to take advantage of interrupts.