@stffrdhrn 2021-04-29
This looks good to me, I was planning on doing a similar patch, is the driver now using multiblock transfers too?  Do you have any performance figures compared to before?