@stffrdhrn 2021-04-29
> is the driver now using multiblock transfers too?

To answer my own question, yes, the statement to limit the block count with `mmc->max_blk_count = 1` is  gone.