@stffrdhrn 2021-04-29
I tested the performance.  In summary in doesn't seem to be much faster than the current branch.

## Without IRQ patches

```
 ~ 618 kB/s write
 ~ 775 kB/s read
 - data is no-corrupted

32-bit OpenRISC CPU with MMU integrated in a LiteX SoC
# uname -a
Linux buildroot 5.12.0-rc8-00030-g6f903ad53b35 #455 Thu Apr 29 14:54:05 JST 2021 openrisc GNU/Linux
```

## With IRQ patches
```
 ~ 541 kB/s write
 ~ 790 kB/s read
 - data is no-corrupted

# uname -a
Linux buildroot 5.12.0-rc8-00036-gf5ef6cc7932d #456 Thu Apr 29 15:03:41 JST 2021 openrisc GNU/Linux

# cat /proc/interrupts 
           CPU0       
  2:        199  or1200-PIC   2  eth0
  3:       5160  or1200-PIC   3  litex-mmc
```

## Notes
My test is a bit dumb, it just dd's a 10mb file of random data with different block sizes.  The 1024 byte block size was the fastest.

On an older kernel I had before the recent mmc driver fixup everything was very slow and there was data corruption.

This required my litex patch to work  : https://github.com/enjoy-digital/litex/pull/901

## Full logs
```
#### ORIG SINGL BLOCK
 
 ~ 100 kB/s write
 ~ 300 kB/s read
 - data is corrupt
 
# /home/shorne/work/gnu-toolchain/mmcperf.sh 
prepare bs=1024
10240+0 records in
10240+0 records out
10485760 bytes (10 MB, 10 MiB) copied, 15.882 s, 660 kB/s
-- perf test bs=256 count=40960 --
 writes bs=256
  10485760 bytes (10 MB, 10 MiB) copied, 120.089 s, 87.3 kB/s
 reads bs=256
  10485760 bytes (10 MB, 10 MiB) copied, 46.5469 s, 225 kB/s
 data check - BAD
  661eda01c7353a4a9bcaacfc703702c2  /tmp/rnd_data.copy
  5fff8c828f1a01844e8f438815fc0723  /tmp/rnd_data.orig

real    0m8.904s
user    0m4.830s
sys     0m3.330s
-rw-r--r-- 1 root root 10485760 Apr 29 05:21 /tmp/rnd_data.copy
-rw-r--r-- 1 root root 10485760 Apr 29 05:18 /tmp/rnd_data.orig
cleanup

-- perf test bs=512 count=20480 --
 writes bs=512
  10485760 bytes (10 MB, 10 MiB) copied, 109.626 s, 95.7 kB/s
 reads bs=512
  10485760 bytes (10 MB, 10 MiB) copied, 35.3468 s, 297 kB/s
 data check
  6af90a4d410b0d3525ad81e0ae30815b  /tmp/rnd_data.copy
  5fff8c828f1a01844e8f438815fc0723  /tmp/rnd_data.orig

real    0m8.209s
user    0m4.870s
sys     0m3.340s
-rw-r--r-- 1 root root 10485760 Apr 29 05:23 /tmp/rnd_data.copy
-rw-r--r-- 1 root root 10485760 Apr 29 05:18 /tmp/rnd_data.orig
cleanup

-- perf test bs=1024 count=10240 --
 writes bs=1024
  10485760 bytes (10 MB, 10 MiB) copied, 103.792 s, 101 kB/s
 reads bs=1024
  10485760 bytes (10 MB, 10 MiB) copied, 29.73 s, 353 kB/s
 data check
  74e9f5b82fd7e1211ed6ec0973b0ed81  /tmp/rnd_data.copy
  5fff8c828f1a01844e8f438815fc0723  /tmp/rnd_data.orig

real    0m13.855s
user    0m9.240s
sys     0m4.060s

#### LITEX REBASE 2021-04-29

 ~ 618 kB/s write
 ~ 775 kB/s read
 - data is no-corrupted

32-bit OpenRISC CPU with MMU integrated in a LiteX SoC
# uname -a
Linux buildroot 5.12.0-rc8-00030-g6f903ad53b35 #455 Thu Apr 29 14:54:05 JST 2021 openrisc GNU/Linux

# /home/shorne/work/gnu-toolchain/mmcperf.sh 
prepare bs=1024
[  169.530000] random: dd: uninitialized urandom read (1024 bytes read)
[  169.530000] random: dd: uninitialized urandom read (1024 bytes read)
[  169.550000] random: dd: uninitialized urandom read (1024 bytes read)
10240+0 records in
10240+0 records out
10485760 bytes (10 MB, 10 MiB) copied, 15.7166 s, 667 kB/s
-- perf test bs=256 count=40960 --
writes bs=256
40960+0 records in
40960+0 records out
10485760 bytes (10 MB, 10 MiB) copied, 34.1702 s, 307 kB/s
reads bs=256
40960+0 records in
40960+0 records out
10485760 bytes (10 MB, 10 MiB) copied, 29.973 s, 350 kB/s
data check
e884a17aa08d31104f9287ab77664958  /tmp/rnd_data.copy
e884a17aa08d31104f9287ab77664958  /tmp/rnd_data.orig

real    0m9.000s
user    0m4.800s
sys     0m4.200s
-rw-r--r-- 1 root root 10485760 Apr 29 06:00 /tmp/rnd_data.copy
-rw-r--r-- 1 root root 10485760 Apr 29 05:59 /tmp/rnd_data.orig
cleanup
-- perf test bs=512 count=20480 --
writes bs=512
20480+0 records in
20480+0 records out
10485760 bytes (10 MB, 10 MiB) copied, 23.705 s, 442 kB/s
reads bs=512
20480+0 records in
20480+0 records out
10485760 bytes (10 MB, 10 MiB) copied, 18.5856 s, 564 kB/s
data check
e884a17aa08d31104f9287ab77664958  /tmp/rnd_data.copy
e884a17aa08d31104f9287ab77664958  /tmp/rnd_data.orig

real    0m8.346s
user    0m4.830s
sys     0m3.490s
-rw-r--r-- 1 root root 10485760 Apr 29 06:01 /tmp/rnd_data.copy
-rw-r--r-- 1 root root 10485760 Apr 29 05:59 /tmp/rnd_data.orig
cleanup
-- perf test bs=1024 count=10240 --
writes bs=1024
10240+0 records in
10240+0 records out
10485760 bytes (10 MB, 10 MiB) copied, 16.9808 s, 618 kB/s
reads bs=1024
10240+0 records in
10240+0 records out
10485760 bytes (10 MB, 10 MiB) copied, 13.537 s, 775 kB/s
data check
e884a17aa08d31104f9287ab77664958  /tmp/rnd_data.copy
e884a17aa08d31104f9287ab77664958  /tmp/rnd_data.orig

real    0m8.442s
user    0m4.800s
sys     0m3.640s
-rw-r--r-- 1 root root 10485760 Apr 29 06:02 /tmp/rnd_data.copy
-rw-r--r-- 1 root root 10485760 Apr 29 05:59 /tmp/rnd_data.orig
cleanup


#### PAULUSMACK PATCHES


 ~ 541 kB/s write
 ~ 790 kB/s read
 - data is no-corrupted

# uname -a
Linux buildroot 5.12.0-rc8-00036-gf5ef6cc7932d #456 Thu Apr 29 15:03:41 JST 2021 openrisc GNU/Linux

# cat /proc/interrupts 
           CPU0       
  2:        199  or1200-PIC   2  eth0
  3:       5160  or1200-PIC   3  litex-mmc
  
# /home/shorne/work/gnu-toolchain/mmcperf.sh 
prepare bs=1024
10240+0 records in
10240+0 records out
10485760 bytes (10 MB, 10 MiB) copied, 15.6314 s, 671 kB/s
-- perf test bs=256 count=40960 --
writes bs=256
40960+0 records in
40960+0 records out
10485760 bytes (10 MB, 10 MiB) copied, 35.1106 s, 299 kB/s
reads bs=256
40960+0 records in
40960+0 records out
10485760 bytes (10 MB, 10 MiB) copied, 29.6651 s, 353 kB/s
data check
5ada256344120e8bc0b7c0fcc33a4608  /tmp/rnd_data.copy
5ada256344120e8bc0b7c0fcc33a4608  /tmp/rnd_data.orig

real    0m8.086s
user    0m4.820s
sys     0m3.250s
-rw-r--r-- 1 root root 10485760 Apr 29 08:35 /tmp/rnd_data.copy
-rw-r--r-- 1 root root 10485760 Apr 29 08:34 /tmp/rnd_data.orig
cleanup
-- perf test bs=512 count=20480 --
writes bs=512
[  389.340000] random: crng init done
[  389.340000] random: 6 urandom warning(s) missed due to ratelimiting
20480+0 records in
20480+0 records out
10485760 bytes (10 MB, 10 MiB) copied, 24.2072 s, 433 kB/s
reads bs=512
20480+0 records in
20480+0 records out
10485760 bytes (10 MB, 10 MiB) copied, 19.4257 s, 540 kB/s
data check
5ada256344120e8bc0b7c0fcc33a4608  /tmp/rnd_data.copy
5ada256344120e8bc0b7c0fcc33a4608  /tmp/rnd_data.orig

real    0m8.209s
user    0m4.810s
sys     0m3.400s
-rw-r--r-- 1 root root 10485760 Apr 29 08:36 /tmp/rnd_data.copy
-rw-r--r-- 1 root root 10485760 Apr 29 08:34 /tmp/rnd_data.orig
cleanup
-- perf test bs=1024 count=10240 --
writes bs=1024
10240+0 records in
10240+0 records out
10485760 bytes (10 MB, 10 MiB) copied, 19.3757 s, 541 kB/s
reads bs=1024
10240+0 records in
10240+0 records out
10485760 bytes (10 MB, 10 MiB) copied, 13.2775 s, 790 kB/s
data check
5ada256344120e8bc0b7c0fcc33a4608  /tmp/rnd_data.copy
5ada256344120e8bc0b7c0fcc33a4608  /tmp/rnd_data.orig

real    0m8.365s
user    0m4.800s
sys     0m3.570s
-rw-r--r-- 1 root root 10485760 Apr 29 08:37 /tmp/rnd_data.copy
-rw-r--r-- 1 root root 10485760 Apr 29 08:34 /tmp/rnd_data.orig
cleanup
```